move to specific file for text client
